# docs/design/Endpoint/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Sparr Admin Endpoints
  version: 1.0.0
  description: |
    Admin endpoint specification. Corresponds to the Admin Endpoint definition in `docs/architecture/components.md`.

    ## Authentication

    Auth routes (sign-in / sign-out / OAuth callbacks) are provided by the Auth adapter at
    `app/api/auth/[...auth]` (framework-managed via `providers.auth.handlers` delegation).
    The concrete URL scheme is adapter-defined and is not specified in this document.

    Business endpoints assume an authenticated principal via `getPrincipal()` and do not own
    sign-in or sign-out flows. `role=admin` is enforced per-route via `requireAdmin()` —
    not by individual endpoint sign-in logic.

    CSRF protection relies on the SameSite cookie policy (Auth.js-independent);
    no X-CSRF-Token header is required.

    All /api/admin/** paths require role=admin enforced per-route via requireAdmin().
    Operations touch ACCOUNT / SCENARIO / PERSONA / RESULT / SETTING via the Store.
    Player endpoints (`/api/player/**`) are documented in `./openapi.yaml`.
servers:
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /api/admin/account:
    get:
      tags: [Account]
      summary: Get admin's own account settings
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Retrieve account ID via getPrincipal()
        4. Call Store.accounts.get(id)
        5. Return AccountProfile (id / email / username / role) with 200

      operationId: getAdminAccount
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Account settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Account]
      summary: Update admin's own account settings
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Accept AccountUpdateRequest (username)
        4. Call Store.accounts.update(id, fields)
        5. Return updated AccountProfile with 200
      operationId: postAdminAccount
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdateRequest"
      responses:
        "200":
          description: Updated account settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ──────────────────────────────────────────
  # Aggregate counts and recent results from the Store.
...
  /api/admin/scenarios:
    get:
      tags: [Scenarios]
      summary: List scenarios
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Call Store.scenarios.list()
        4. Return AdminScenarioSummary[] with 200
      operationId: getAdminScenarios
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Scenario list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminScenarioSummary"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Store retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Scenarios]
      summary: Create a scenario
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Accept AdminScenarioCreateRequest (scenario body + optional personas[])
        4. Assign a scenario id: use the provided id, or generate one (uid()) if omitted
        5. Call Store.scenarios.create(scenario)
        6. If personas[] provided, assign an id (uid()) to any persona missing one,
           then call Store.personas.replaceAll(id, personas)
        7. Return the created AdminScenarioDetail with 201
      operationId: postAdminScenarios
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminScenarioCreateRequest"
      responses:
        "201":
          description: Scenario created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminScenarioDetail"
        "400":
          description: Validation error (missing required field, invalid roles/documents/rubricItems, etc.)

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/scenarios/{id}:
    get:
      tags: [Scenarios]
      summary: Get scenario detail (including personas)
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Call Store.scenarios.get(id) (404 if not found)
        4. Call Store.personas.findByScenarioId(id)
        5. Return AdminScenarioDetail (scenario + personas[]) with 200
      operationId: getAdminScenarioById
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Scenario ID
      responses:
        "200":
          description: Scenario detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminScenarioDetail"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Store retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Scenarios]
      summary: Update a scenario
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Accept AdminScenarioUpdateRequest (partial Scenario field updates + optional personas[])
        4. Call Store.scenarios.update(id, fields) (404 if not found)
        5. If personas[] provided, assign an id (uid()) to any persona missing one
           (existing personas keep their id if included), then call
           Store.personas.replaceAll(id, personas)
        6. Return updated AdminScenarioDetail with 200
      operationId: postAdminScenarioById

      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Scenario ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminScenarioUpdateRequest"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminScenarioDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Scenarios]
      summary: Delete a scenario
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Call Store.scenarios.delete(id) (404 if not found)
        4. Return 200


        - If a scenario referenced by existing RESULTs is deleted, the RESULTs remain but scenarioId becomes orphaned
      operationId: deleteAdminScenarioById
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Scenario ID
      responses:
        "200":
          description: Deletion successful

        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: |
        Session cookie set by the Auth adapter on sign-in. Sent automatically by the browser
        for same-origin requests. CSRF protection relies on the SameSite cookie policy —
        no separate X-CSRF-Token header is required.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth / Account ────────────────────────────────────
...
    AccountProfile:
      type: object
      required: [id, role, email, username]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [player, admin]
    AccountUpdateRequest:
      type: object
      properties:
        username:
          type: string

    # ── Dashboard ─────────────────────────────────────────
...
    AdminPersona:
      type: object
      required: [id, scenarioId, characterPrompt, voiceCode, docToolEnabled]
      properties:
        id:
          type: string
        scenarioId:
          type: string
        characterPrompt:
          type: string
        voiceCode:
          type: string
        docToolEnabled:
          type: boolean
    AdminPersonaInput:
      type: object
      required: [characterPrompt, voiceCode, docToolEnabled]
      properties:
        id:
          type: string
          description: |
            Omit to auto-generate a new persona ID (uid()). When updating a scenario,
            include the existing id to keep that persona's id stable; personas omitted
            from the array are removed (replaceAll semantics).
        characterPrompt:
          type: string
        voiceCode:
          type: string
        docToolEnabled:
          type: boolean
    AdminScenarioSummary:
      type: object
      required: [id, title, challengePrompt]
      properties:
        id:
          type: string
        title:
          type: string
        challengePrompt:
          type: string
    AdminScenarioDetail:
      type: object
      required: [id, basePrompt, challengePrompt, documentsPrompt, rubricPrompt, personas]
      properties:
        id:
          type: string
        basePrompt:
          type: string
        challengePrompt:
          type: string
        documentsPrompt:
          type: string
        rubricPrompt:
          type: string
        personas:
          type: array
          items:
            $ref: "#/components/schemas/AdminPersona"
    AdminScenarioCreateRequest:
      type: object
      required: [basePrompt, challengePrompt, documentsPrompt, rubricPrompt]
      properties:
        id:
          type: string
          description: Omit to auto-generate a new scenario ID (uid()).
        basePrompt:
          type: string
        challengePrompt:
          type: string
        documentsPrompt:
          type: string
        rubricPrompt:
          type: string
        personas:
          type: array
          items:
            $ref: "#/components/schemas/AdminPersonaInput"
    AdminScenarioUpdateRequest:
      type: object
      properties:
        basePrompt:
          type: string
        challengePrompt:
          type: string
        documentsPrompt:
          type: string
        rubricPrompt:
          type: string
        personas:
          type: array
          items:
            $ref: "#/components/schemas/AdminPersonaInput"
...